requirements.txt
albumentations==1.3.1
contourpy==1.1.1
cycler==0.11.0
fonttools==4.42.1
imageio==2.31.3
joblib==1.3.2
json5==0.9.14
kiwisolver==1.4.5
lazy_loader==0.3
matplotlib==3.8.0
networkx==3.1
numpy==1.26.0
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
packaging==23.1
pandas==2.1.0
Pillow==10.0.1
pycocotools==2.0.7
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
qudida==0.0.4
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
shapely==2.0.1
six==1.16.0
threadpoolctl==3.2.0
tifffile==2023.8.30
typing_extensions==4.8.0
tzdata==2023.3
